Create Strings1.java
This code defines a class Strings1 that, when executed, initializes two strings s1 with the value "Computer Science" and s2 with the value "Education". It then extracts a substring from s1 starting from index 9 up to (but not including) index 16, which results in the word "Science". Finally, it prints the extracted substring followed by the string "Education", resulting in the output "Science Education"